完成模型广场的模型展示功能，通过点击广场中的卡片可以跳转到对应的模型的的对话页面，完善了新增会话的功能，修改了原来的错误逻辑

+++ src/main/java/com/zqw/agent_app/model/vo/AgentVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class AgentVO {
 
-    private Long agentId;
+    private int agentId;
     private String agentName;
     private String description;
     private Integer usageCount;
